refactor(handler): move Create, Read, Update, and Delete functions to separate files
