+++ internal/handler/kero_create.go
@@ -1,0 +1,33 @@
+package handler
+
+import (
+	"github.com/kaerubo/kaeruashi/internal/entity"
+	"github.com/labstack/echo/v4"
+	"net/http"
+)
+
+type createKeroRequest struct {
+	Title   string `json:"title"`
+	Content string `json:"content"`
+}
+
+func (h *KeroHandler) CreateKero(c echo.Context) error {
+	var req createKeroRequest
+	if err := c.Bind(&req); err != nil {
+		return echo.NewHTTPError(http.StatusBadRequest, "invalid request")
+	}
+	if req.Title == "" || req.Content == "" {
+		return echo.NewHTTPError(http.StatusBadRequest, "title and content are required")
+	}
+
+	kero := &entity.Kero{
+		Title:   req.Title,
+		Content: req.Content,
+	}
+
+	if err := h.creator.Create(c.Request().Context(), kero); err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	return c.JSON(http.StatusCreated, map[string]string{"id": kero.ID})
+}
